+++ src/layers/models/layersManager.ts
@@ -1,7 +1,15 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { inject, injectable } from 'tsyringe';
 import { Services } from '../../common/constants';
-import { ILogger, ILayerPostRequest, IMapProxyCache, IMapProxyJsonDocument, IMapProxyLayer, IMapProxyConfig, ILayerToBestRequest } from '../../common/interfaces';
+import {
+  ILogger,
+  ILayerPostRequest,
+  IMapProxyCache,
+  IMapProxyJsonDocument,
+  IMapProxyLayer,
+  IMapProxyConfig,
+  ILayerToBestRequest,
+} from '../../common/interfaces';
 import { mockLayer } from '../../common/data/mock/mockLayer';
 import { convertJsonToYaml, convertYamlToJson, replaceYamlFileContent } from '../../common/utils';
 import { ConfilctError } from '../../common/exceptions/http/confilctError';
