+++ tests/unit/layers/models/layersManager.spec.ts
@@ -1,10 +1,11 @@
 import config from 'config';
-import { IMapProxyConfig } from '../../../../src/common/interfaces';
+import { ILayerToBestRequest, IMapProxyConfig } from '../../../../src/common/interfaces';
 import { LayersManager } from '../../../../src/layers/models/layersManager';
 import { ConfilctError } from '../../../../src/common/exceptions/http/confilctError';
 import { mockLayerNameAlreadyExists } from '../../mock/mockLayerNameAlreadyExists';
 import { mockLayerNameIsNotExists } from '../../mock/mockLayerNameIsNotExists';
 import * as utils from '../../../../src/common/utils';
+import { NoContentError } from '../../../../src/common/exceptions/http/noContentError';
 
 let layersManager: LayersManager;
 let convertYamlToJsonstub: jest.SpyInstance;
@@ -47,6 +48,50 @@ describe('layersManager', () => {
     it('should successfully add layer', function () {
       // action
       const action = () => layersManager.addLayer(mockLayerNameIsNotExists);
+      // expectation
+      expect(action).not.toThrow();
+      expect(convertYamlToJsonstub).toHaveBeenCalledTimes(1);
+      expect(convertJsonToYamlStub).toHaveBeenCalledTimes(1);
+      expect(replaceYamlContentStub).toHaveBeenCalledTimes(1);
+    });
+  });
+  describe('#addLayerToBest', () => {
+    it('should reject with no content error due layer name is not exists', function () {
+      // mock
+      const mockLayerNotExistsToBestRequest: ILayerToBestRequest = {
+        layerName: 'layerNameIsNotExists',
+        bestName: 'bestMockName',
+      };
+      // action
+      const action = () => layersManager.addLayerToBest(mockLayerNotExistsToBestRequest);
+      // expectation
+      expect(action).toThrow(NoContentError);
+      expect(convertYamlToJsonstub).toHaveBeenCalledTimes(1);
+      expect(convertJsonToYamlStub).not.toHaveBeenCalled();
+      expect(replaceYamlContentStub).not.toHaveBeenCalled();
+    });
+    it('should reject with no content error due best name is not exists', function () {
+      // mock
+      const mockLayerToBestRequest: ILayerToBestRequest = {
+        layerName: 'mockLayerNameExists',
+        bestName: 'bestNameIsNotExists',
+      };
+      // action
+      const action = () => layersManager.addLayerToBest(mockLayerToBestRequest);
+      // expectation
+      expect(action).toThrow(NoContentError);
+      expect(convertYamlToJsonstub).toHaveBeenCalledTimes(1);
+      expect(convertJsonToYamlStub).not.toHaveBeenCalled();
+      expect(replaceYamlContentStub).not.toHaveBeenCalled();
+    });
+    it('should successfully add layer to best', function () {
+      // mock
+      const mockLayerToBestRequest: ILayerToBestRequest = {
+        layerName: 'mockLayerNameExists',
+        bestName: 'existsBestName',
+      };
+      // action
+      const action = () => layersManager.addLayerToBest(mockLayerToBestRequest);
       // expectation
       expect(action).not.toThrow();
       expect(convertYamlToJsonstub).toHaveBeenCalledTimes(1);
